Update ManualComputation round-tripping to not have round-tripping attrs on the CallOps.
This is needed as sometimes the op has `stablehlo.token` types, and during MLIR<->HLO round-tripping, MLIR type conversion converts this to `mhlo.token` and "accidentally" discards the unregistered attrs (the `frontend_attributes`) - note MLIR has no guarantees about preserving these, so we can't submit a fix to MLIR.

This happens because StableHLO->HLO conversion still does StableHLO->MHLO conversion, but this should be removed soon. To unblock ourselves instead of waiting for StableHLO to remove the intermediate MHLO pass, we can just move the frontend attrs to the GlobalToLocal/LocalToGlobal custom calls.

PiperOrigin-RevId: 745735176